refactor: extract databend-meta-raft-config crate
# Summary

Node configuration, on-disk format versioning and the startup error move
out of `databend-meta-raft-store` into a new leaf crate. Nothing in the
new crate depends on anything of ours except `meta-types`, so it now
compiles in parallel with the rest of the store instead of behind it.

# Details

The crate holds `config`, `snapshot_config`, `data_version`,
`meta_startup_error` and `StateMachineFeature` — the layer that says
where data lives and which format it is in, with no knowledge of how
that data is read or written. Keeping it separate is what stops a future
edit from making the config layer depend on the storage layer again.

Two methods widen from `pub(crate)` to `pub`
(`SnapshotConfig::ensure_snapshot_dir`, `VersionInfo::download_url`), as
does the `version_info` module. They already crossed module boundaries;
the split only makes that visible.

This is the first of six planned extractions out of `raft-store`.
Call sites move from `databend_meta_raft_store::config::…` to
`databend_meta_raft_config::config::…`; no facade is left behind.
