+++ crates/server/raft-config/Cargo.toml
@@ -1,0 +1,34 @@
+[package]
+name = "databend-meta-raft-config"
+description = "Node configuration and on-disk format versioning"
+version = { workspace = true }
+authors = { workspace = true }
+license = { workspace = true }
+publish = { workspace = true }
+edition = { workspace = true }
+
+[dependencies]
+anyerror = { workspace = true }
+databend-meta-runtime-api = { workspace = true }
+databend-meta-types = { workspace = true }
+hostname = { workspace = true }
+log = { workspace = true }
+maplit = { workspace = true }
+openraft = { workspace = true }
+raft-log = { workspace = true }
+rotbl = { workspace = true }
+semver = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+strum = { workspace = true }
+strum_macros = { workspace = true }
+thiserror = { workspace = true }
+
+[dev-dependencies]
+anyhow = { workspace = true }
+pretty_assertions = { workspace = true }
+tempfile = { workspace = true }
+tokio = { workspace = true }
+
+[lints]
+workspace = true
