+++ crates/server/raft-config/src/data_version/mod.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub(crate) mod version_info;
+pub mod version_info;
 
 use std::fmt;
