+++ crates/server/raft-store/src/header.rs
@@ -14,12 +14,11 @@
 
 use std::fmt;
 
+use databend_meta_raft_config::data_version::DATA_VERSION;
+use databend_meta_raft_config::data_version::DataVersion;
 use databend_meta_sled_store::IVec;
 use databend_meta_sled_store::SledBytesError;
 use databend_meta_sled_store::SledSerde;
-
-use crate::data_version::DATA_VERSION;
-use crate::data_version::DataVersion;
 
 #[derive(
     Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, serde::Serialize, serde::Deserialize,
