+++ crates/server/raft-store/src/ondisk/ondisk_test.rs
@@ -21,6 +21,8 @@ use std::sync::Arc;
 use std::sync::Mutex;
 use std::sync::MutexGuard;
 
+use databend_meta_raft_config::config::RaftConfig;
+use databend_meta_raft_config::data_version::DataVersion;
 use databend_meta_runtime_api::TokioRuntime;
 use databend_meta_sled_store::SledTree;
 use databend_meta_sled_store::drop_sled_db;
@@ -33,8 +35,6 @@ use databend_meta_types::raft_types::new_log_id;
 use pretty_assertions::assert_eq;
 use tempfile::TempDir;
 
-use crate::config::RaftConfig;
-use crate::data_version::DataVersion;
 use crate::header::Header;
 use crate::log_store::Cw;
 use crate::log_store::LogStoreMeta;
