+++ crates/server/service/Cargo.toml
@@ -18,6 +18,7 @@ chrono = { workspace = true }
 databend-base = { workspace = true }
 databend-meta-client = { workspace = true }
 databend-meta-kvapi = { workspace = true }
+databend-meta-raft-config = { workspace = true }
 databend-meta-raft-store = { workspace = true }
 databend-meta-runtime-api = { workspace = true }
 databend-meta-sled-store = { workspace = true }
