+++ crates/server/service/src/configs/inner.rs
@@ -14,8 +14,8 @@
 
 use std::net::SocketAddr;
 
-use databend_meta_raft_store::MetaStartupError;
-use databend_meta_raft_store::config::RaftConfig;
+use databend_meta_raft_config::MetaStartupError;
+use databend_meta_raft_config::config::RaftConfig;
 use databend_meta_types::node::Node;
 
 /// TLS configuration for server endpoints.
