+++ crates/server/service/src/meta_node/meta_node_builder.rs
@@ -15,7 +15,7 @@
 use std::sync::Arc;
 use std::sync::atomic::AtomicI32;
 
-use databend_meta_raft_store::MetaStartupError;
+use databend_meta_raft_config::MetaStartupError;
 use databend_meta_runtime_api::SpawnApi;
 use databend_meta_types::Endpoint;
 use databend_meta_types::raft_types::NodeId;
